requirements.txt
openai
groq
serpapi
tiktoken
python-dotenv
newsapi-python
